requirements.txt
requests==2.26.0
pandas==1.3.3
PyPDF2==1.26.0
pdfplumber==0.5.28
beautifulsoup4==4.9.3
matplotlib==3.4.3
psycopg2-binary==2.9.3